+++ cmd/generate/generate.go
@@ -6,16 +6,15 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/open-sauced/pizza-cli/cmd/generate/codeowners"
+	"github.com/open-sauced/pizza-cli/cmd/generate/config"
 )
 
-const generateLongDesc string = `WARNING: Proof of concept feature.
-
-XXX`
+const generateLongDesc string = `The 'generate' command provides tools to automate the creation of important project documentation and derive insights from your codebase.`
 
 func NewGenerateCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "generate [subcommand] [flags]",
-		Short: "Generates something",
+		Short: "Generates documentation and insights from your codebase",
 		Long:  generateLongDesc,
 		Args: func(_ *cobra.Command, args []string) error {
 			if len(args) != 1 {
@@ -28,6 +27,7 @@ func NewGenerateCommand() *cobra.Command {
 	}
 
 	cmd.AddCommand(codeowners.NewCodeownersCommand())
+	cmd.AddCommand(config.NewConfigCommand())
 
 	return cmd
 }